pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "peakrdl-cpp"
version = "0.1.0"
description = "PeakRDL exporter that generates C++ register maps"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
dependencies = [
  "peakrdl>=1.5.0",
  "systemrdl-compiler>=1.32.0",
]

[project.entry-points."peakrdl.exporters"]
cpp = "peakrdl_cpp.__peakrdl__:Exporter"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
